python/requirements.txt
ffmpeg-python==0.1.17
future==0.17.1
numpy==1.16.2
pep8==1.7.1
Pillow==5.4.1
pycodestyle==2.5.0
scikit-learn==0.20.3
scipy==1.2.1
